fix(formtemplates): fixed requested changes
